% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fra_arrivals_departures.R
\name{distinct_fra_df}
\alias{distinct_fra_df}
\title{distinct_fra_df}
\usage{
distinct_fra_df(fra_df)
}
\arguments{
\item{df}{Data frame containing either arrival or departure data}
}
\value{
updated data frame with the newly read data; duplicates removed; sorted by id
}
\description{
Remove duplicate lines from flight data frame
}
\details{
Data on the web site is updated every couple of minutes; getting new time stamps:
a lu stamp for the last change of data in the FRA database, and a timestamp from
the Rfraflights query. Ignore these time stamps and remove all duplicate data,
keeping only the first occurrence of modified data.

Aussumes that id is an unique identifier for ONE flight on a certain day.

Basically just a helper function but exported for use in data sanitation.
}
