% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arrivals_departures.R
\name{get_arrivals_df}
\alias{get_arrivals_df}
\title{get_fra_data}
\usage{
get_arrivals_df(h_back = 48, h_forward = 120)
}
\description{
R Wrapper for querying the JSON-API on frankfurt-airport.com
}
\author{
Jan Eggers <jan.eggers@hr.de>
}
