% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fra_data.R
\name{get_fra}
\alias{get_fra}
\title{get_fra}
\usage{
get_fra(
  flighttype = "arrivals",
  from = now() - hours(48),
  to = NA,
  pages = 1,
  perpage = 50
)
}
\arguments{
\item{flighttype}{"arrivals" or "departures"}

\item{from}{Point of time for querying: Returns flights with a scheduled date equal to or higher than time. Time must be given as UTC - please note that all return times are given as local times, with the time difference to UTC.}

\item{to}{Upper sched datetime limit for queried flights}

\item{pages}{Number of pages to be queried - used only if no `to` parameter is given}

\item{perpage}{Entries per page (max. 50) - use only if querying by number of entries}
}
\value{
Data frame - see README.md for variables
}
\description{
`get_fra` queries flight information from FRA
}
\details{
Query flights, ordered by scheduled departure, starting at a point in time (up to to 48 hours in the past, or in the future), up to an end point, or a number of pages. Returns a data frame with one flight per line.
}
\examples{
get_fra("arrivals", from=now(), pages=1)
get_fra("departures", from=now()-hours(24), to=now()+hours(24))
get_fra("departures", from=now(), pages=10) # gets next 500 flights

}
