% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fra_arrivals_departures.R
\name{get_fra_departures_df}
\alias{get_fra_departures_df}
\title{get_fra_departures_df}
\usage{
get_fra_departures_df(h_back = 48, h_forward = 120)
}
\arguments{
\item{h_back}{Number of hours to look back from now (max. 48; larger values are clipped). Negative values are converted.}

\item{h_forward}{Number of hours to look forward}
}
\value{
Data frame containing flight data, ordered by scheduled datetime (see project README for details)
}
\description{
Read departure data into a data frame
}
\details{
Using the get_fra_data function, the FRA website is queried for scheduled departures and their actual status, looking back the given number of hours into the past schedule, and the number of hours into the future, as by the scheduled date/time of the flight.
}
\examples{
get_fra_departures_df()
get_fra_departures_df(h_back = -24, h_forward = 24)
#NOT RUN:
get_departures_df(1,2400)

}
