% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fra_data.R
\name{get_fra_flight}
\alias{get_fra_flight}
\title{get_fra_flight}
\usage{
get_fra_flight(id)
}
\arguments{
\item{id}{e.g. "a20220712lh1284" consists of:
- "a"rrival or "d"eparture
- the scheduled date, formatted YYYYMMDD
- airline and flight code}
}
\value{
data frame with basically the same parameters as in the paginated data
}
\description{
Low-level function to query one single flight via ID
}
