% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fra_data.R
\name{get_fra_page}
\alias{get_fra_page}
\title{get_fra_page}
\usage{
get_fra_page(flighttype, time, perpage = 50, page = 1)
}
\arguments{
\item{flighttype}{"arrivals" or "departures"}

\item{time}{Point of time for querying: Returns flights with a scheduled date equal to or higher than time. Time must be given as UTC - please note that all return times are given as local times, with the time difference to UTC.}

\item{perpage}{Entries per page (max. 50)}

\item{page}{Number of the page to be queried - reading all entries with a scheduled date time from ´time´ onwards may return a potentially unlimited number of flights. They are sorted by ascending schedule datetime and paginated - i.e. with perpage=50, page=1 returns entries 1 to 50, page=2 returns 51 to 100, and so on.}
}
\value{
Data frame
}
\description{
`get_fra_page` queries one page of flight information from FRA
}
\details{
Query one page of flight information (max. 50 entries) by scheduled date and time. Query time may be up to 48 hours in the past, or in the future. Returns a data frame with one flight per line.
}
\examples{
get_fra_page("arrivals", time=now(), page=1)
get_fra_page("departures", time=now()+hours(24), page = 5, perpage = 10)

}
