% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fra_arrivals_departures.R
\name{update_fra_df}
\alias{update_fra_df}
\title{update_fra_df}
\usage{
update_fra_df(df, h_back = 48, h_forward = 120)
}
\arguments{
\item{df}{Data frame containing either arrival or departure data}

\item{h_back}{hours to look backward (max 48)}

\item{h_forward}{hours to look forward}
}
\value{
updated data frame with the newly read data; duplicates removed
}
\description{
Reads arrivals/departures and appends to existing data, removing duplicates
}
