% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fra_arrivals_departures.R
\name{update_fra_log}
\alias{update_fra_log}
\title{update_fra_log}
\usage{
update_fra_log(path = "./fra_log/", flighttype = "arrivals")
}
\arguments{
\item{path}{path to log files (e.g. "/path/")}

\item{flighttype}{"arrivals" or "departures" (or short: "a" or "d")}
}
\value{
TRUE if valid
}
\description{
Reads FRA flight data and appends to .RDS and .csv files
}
\details{
Assumes that there is a file "arrivals.RDS" (or "departures.RDS") in the current directory containing a data frame with all previously recorded data. Updates the file with any flight data that has beek changed in the meantime, keeping the original record so that you can calculate the delay and change times. Also writes a arrivals.csv (or departures.csv)
}
